Removed a bunch of dead code and fixed bug in quadrotor collision detection.

+++ includes/worlds.hpp
@@ -156,12 +156,12 @@ public:
 	}
 
 	void showCollisionCheck(const state& s, bool collision, int collision_hit_group, int collision_free_group) {
-		int x_pos = s[0];
-		int y_pos = s[1];
+		double x_pos = s[0];
+		double y_pos = s[1];
 #if POSITION_DIM == 3
-		int z_pos = s[2];
+		double z_pos = s[2];
 #else
-		int z_pos = 0;
+		double z_pos = 0;
 #endif
 		if (collision) {
 #ifdef SHOW_COLLISIONS
